Functionalities Done

+++ backend/app/Controllers/post/post_service.ts
@@ -1,5 +1,7 @@
 import { DateTime } from "luxon";
 import PostQuery from "./post_query.js";
+import Post from "#models/post";
+import { Exception } from "@adonisjs/core/exceptions";
 export default class PostService{
     private postQuery:PostQuery
     constructor(){
@@ -8,11 +10,25 @@ export default class PostService{
     public async CreatePost(data:{content:string,userId:number}){
         return await this.postQuery.createPost(data.content,data.userId)
     }
-    public async DeletePost(data:{postId:number}){
-        return await this.postQuery.deletePost(data.postId);
+    public async DeletePost(data:{postId:number},userId:number){
+        const pid=await Post.query().where('id',data.postId);
+        console.log(pid);
+        if(pid[0].$attributes.userId===userId){
+            return await this.postQuery.deletePost(data.postId);
+        }
+        else {
+            throw new Exception("Error While Deleting")
+        }
     }
-    public async UpdatePost(data:{postId:number,content:string,userId:number}){
-        return await this.postQuery.updatePost(data.postId,data.content,data.userId)
+    public async UpdatePost(data:{postId:number,content:string},userId:number){
+        const pid=await Post.query().where('id',data.postId);
+        console.log(pid);
+        if(pid[0].$attributes.userId===userId){
+            return await this.postQuery.updatePost(data.postId,data.content,userId)
+        }
+        else {
+            throw new Exception("Error While Updating")
+        }
     }
     public async GetLike(data:{postId:number}){
         return await this.postQuery.getLike(data.postId)
